+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Remove whitespace from C/C++ lines in the Assembly tab.
 * Allow assembly code to have its own font format. See Config->Editor->AssemblyText.
 * Add Nexti and Stepi toolbar buttons if Assembly tab is shown.
+* Add PC, SP, FLAGS status bar to Assembly tab.
 
 ## [1.13] - 2022-12-02